Refactor MCP servers page
- Simplified the structure by extracting components for server card display and server addition dialog.
- Introduced a new server card list component to manage server display logic.
- Enhanced server fetching and connection testing with hooks for better state management using SWR and our zustand store
- Updated action handlers for server addition, deletion, and toggling to streamline operations.
- Improved error handling and user feedback through toast notifications.
