+++ lib/hooks/index.ts
@@ -1,0 +1,2 @@
+export { useServerInfo, usePrefetchServerInfo } from './use-mcp-server-info';
+export { useServerTest, useTestServerConnection } from './use-mcp-server-test';
